minor, finalizing changes

+++ Reference Sequence Alignment.py
@@ -24,6 +24,8 @@ RAW_fileNameList = []
 for f in ["control", "treated"]:
     for s in ["1", "2"]:
         RAW_fileNameList.append("RAW_DATA/tRNA_human_" + f + "_" + s + ".fasta")
+RAW_fileNameList.remove("RAW_DATA/tRNA_human_control_1.fasta")
+
 #getting the list of seed sequences
 SEQ_LIST = []
 SEQ_DICT = {}
@@ -31,9 +33,12 @@ for fileName in fileNameList:
     handle = open(fileName, "r")
     for line in handle:
         if line[0] != ">":
-            if line[:-1] != "":
-                SEQ_LIST.append(line[:-1])
-                SEQ_DICT[line[:-1]] = []
+            seq = line[:-1]
+            if "C" in seq:
+                while seq[-1] != "A" and seq[-1] != "C":
+                    seq = seq[:-1]
+                SEQ_LIST.append(seq)
+                SEQ_DICT[seq] = []
     handle.close()
 
 UNIQUE = []
@@ -49,7 +54,12 @@ toAlignLater = {}
 
 def checkSeq(seq1, seq2):
     if seq1[-1] == "C":
+##        print("here")
+##        print(seq1)
+##        print(seq2)
+##        print(seq2[-1])
         seq2 = seq2[:-1]
+##        print(seq2)
     if len(seq1) > len(seq2):
         seq1 = seq1[len(seq1) - len(seq2):]
     else:
@@ -58,10 +68,12 @@ def checkSeq(seq1, seq2):
     for i in range(len(seq1)):
         if seq1[i] != seq2[i]:
             mismatch += 1
-            if mismatch > 2:
-                return False
-    return True
+            if mismatch > 3:
+                return (False, 0)
+    return (True, mismatch)
 
+checked = 0
+accepted = 0
 for fileName in RAW_fileNameList:
     print(fileName)
     i = 0
@@ -69,30 +81,46 @@ for fileName in RAW_fileNameList:
     IDline = ""
     for line in handle:
         if line[0] == ">":
-            i += 1
-            if i % 10000 == 0:
-                print(i/1000)
+            if checked % 10000 == 0:
+                print(checked, accepted)
+##            i += 1
+##            if i % 10000 == 0:
+##                print(i/1000)
             IDline = line
         else:
             seq = line[:-1]
             if seq != "":
                 matches = []
-                for SEQ in SEQ_LIST:
-                    if checkSeq(seq, SEQ):
-                        matches.append(SEQ)
-                if len(matches) == 1:
-                    SEQ_DICT[matches[0]].append(Sequence(seq, IDline))
-                elif len(matches) > 1:
-                    stringMatches = ",".join(matches)
-                    if stringMatches in toAlignLater:
-                        toAlignLater[stringMatches].append(Sequence(seq, IDline))
-                    else:
-                        toAlignLater[stringMatches] = [Sequence(seq, IDline)]
+                checked += 1
+                if "C" in seq:
+                    while seq[-1] != "A" and seq[-1] != "C":
+                        seq = seq[:-1]
+                    lowest_misses = 4
+                    for SEQ in SEQ_LIST:
+                        check_output = checkSeq(seq, SEQ)
+                        if check_output[0]:
+                            if check_output[1] < lowest_misses:
+                                matches = []
+                                matches.append(SEQ)
+                                lowest_misses = check_output[1]
+                            elif check_output[1] == lowest_misses:
+                                matches.append(SEQ)
+                    if len(matches) == 1:
+                        accepted += 1
+                        SEQ_DICT[matches[0]].append(Sequence(seq, IDline))
+                    elif len(matches) > 1:
+                        accepted += 1
+                        stringMatches = ",".join(matches)
+                        if stringMatches in toAlignLater:
+                            toAlignLater[stringMatches].append(Sequence(seq, IDline))
+                        else:
+##                            print(len(toAlignLater))
+                            toAlignLater[stringMatches] = [Sequence(seq, IDline)]
     handle.close()
 print(len(toAlignLater))
 for stringmatches in toAlignLater:
     matches = stringmatches.split(",")
-    print(matches)
+##    print(matches)
     seq_list = toAlignLater[stringmatches]
     numSeqeunces = len(seq_list)
     counts = []
@@ -122,14 +150,18 @@ for stringmatches in toAlignLater:
                 toAdd = matches[random.randint(0, len(matches) - 1)]
             SEQ_DICT[toAdd].append(seq_list[num_seq])
             num_seq += 1
-
+output = ""
 for SEQ in SEQ_LIST:
     keep = True
     if not keepZeroes:
         if len(SEQ_DICT[SEQ]) == 0:
             keep = False
     if keep:
-        output = open("sequence_alignments/" + SEQ + ".txt", "w")
+        if os.path.exists("sequence_alignments/" + SEQ + ".txt"):
+            output = open("sequence_alignments/" + SEQ + ".txt", "a")
+            print("Appending...")
+        else:
+            output = open("sequence_alignments/" + SEQ + ".txt", "w")
         for elem in SEQ_DICT[SEQ]:
             output.write(elem.getID() + elem.getSeq() + "\n")
         output.close()
